+++ tests/ignite/metrics/regression/test_r2_score.py
@@ -34,8 +34,8 @@ def test_r2_score(available_device):
 
     m = R2Score(device=available_device)
     assert m._device == torch.device(available_device)
-    y_pred = torch.from_numpy(np_y_pred)
-    y = torch.from_numpy(np_y)
+    y_pred = torch.from_numpy(np_y_pred).to(dtype=torch.float32)
+    y = torch.from_numpy(np_y).to(dtype=torch.float32)
 
     m.reset()
     m.update((y_pred, y))
@@ -52,8 +52,8 @@ def test_r2_score_2(available_device):
 
     m = R2Score(device=available_device)
     assert m._device == torch.device(available_device)
-    y_pred = torch.from_numpy(np_y_pred)
-    y = torch.from_numpy(np_y)
+    y_pred = torch.from_numpy(np_y_pred).to(dtype=torch.float32)
+    y = torch.from_numpy(np_y).to(dtype=torch.float32)
 
     m.reset()
     batch_size = 16
@@ -78,7 +78,9 @@ def test_integration_r2_score(available_device):
         idx = (engine.state.iteration - 1) * batch_size
         y_true_batch = np_y[idx : idx + batch_size]
         y_pred_batch = np_y_pred[idx : idx + batch_size]
-        return torch.from_numpy(y_pred_batch), torch.from_numpy(y_true_batch)
+        return torch.from_numpy(y_pred_batch).to(dtype=torch.float32), torch.from_numpy(y_true_batch).to(
+            dtype=torch.float32
+        )
 
     engine = Engine(update_fn)
